Fixed color flickering when using modifed ImGradient (#238)
* Fixed color flickering when using modifed ImGradient

* Changed curves to work in multiple incles, Andreu's cooking.
