+++ SobrassadaEngine/Libs/ImGui/imgui_color_gradient.cpp
@@ -50,7 +50,7 @@ void ImGradient::getColorAt(float position, float* color) const
 {
     position      = ImClamp(position, 0.0f, 1.0f);
     int cachePos  = (int)(position * 255);
-    cachePos     *= 3;
+    cachePos     *= 4;
     color[0]      = m_cachedValues[cachePos + 0];
     color[1]      = m_cachedValues[cachePos + 1];
     color[2]      = m_cachedValues[cachePos + 2];
